Fixes Bluebird theme with root user GTK theme
Copies the GTK3 symlink and GTK2 config file from the ubuntu directory to the
root home directory. The duplication of the GTK2 config file is not ideal but
is not a big deal.

This ensures Rescuezilla application runs with the attractive Bluebird theme,
and the lxterminal application runs as root with intended theming (blue menu
bar).

This is fixes part of task #48 and is somewhat relevant to task #52.