Adds Apple Filesystem (APFS) support to frontend
Adds ability for Rescuezilla frontend to use partclone to make filesystem-aware
backup/restore of Apple Filesystem (APFS), which is the default filesystem on
macOS High Sierra and newer.

IMPORTANT: In testing, partclone v0.3.13 as available on Ubuntu 20.04 Focal has
immature APFS support exiting with "bitmap error" when using a real APFS
filesystem (log excerpt in #65). Until there is a focal-backports (or
focal-proposed) version of partclone (perhaps v0.3.14 [2]), this commit should
not be merged.

A partclone.apfs executable is available on Ubuntu 20.04 (Rescuezilla 64-bit)
but not on Ubuntu 18.04 (Rescuezilla 32-bit).

[1] https://launchpad.net/ubuntu/+source/partclone

[2] https://github.com/Thomas-Tsai/partclone