Clarifies Canonical-pkg build env requirement
Rewords the reason behind requiring a Canonical package environment for
building a Rescuezilla image with working Secure Boot to make it clearer that
the Microsoft-signed shim does *not* contain Canonical certificates.

My understanding is GRUB is signed by Canonical using a certificate chain where
Microsoft is the authority, but the shim contains no Canonical certificates.