Copies all of post-MBR gap, not just 32KiB
Calculates the amount of bytes until the first partition. Modern disks have
filesystems aligned to 1MiB (sector 2048) rather than starting after a 32KiB
gap.

On MBR disks, this "post-MBR gap" (the gap between the end of the Master Boot
Record at 512 bytes offset and hte first partition) contains the GRUB2 core.img
executable (called stage 1.5 on GRUB legacy). This bootloader is historically
smaller than 32KiB, but through embedding different modules and 1MiB partition
alignment may be larger.

It's important to copy the entirety of the post-MBR gap to ensure that larger
GRUB core.img are not truncated on MBR disks, which breaks booting with GRUB.

Finally, this gap is copied whether or not the disk uses GPT (the GPT table
itself resides in the gap)

Fixes #50
Fixes #64