Installs different packages on 32bit vs 64bit
Adds the ability to install different packages on the 32-bit version of
Rescuezilla compared to the 64-bit version. This is important given the 64-bit
version is based on Ubuntu 20.04 (Focal) as of writing, and the 32-bit version
is based on Ubuntu 18.04 (Bionic), given there are packages available on 20.04
which aren't available on 18.04 (and vice versa). This change only adds the
ability to do this, as of this commit the exact package lists are unchanged.

Also adds an `apt-get autoclean` which should keep the pkg.cache small when as
packages in the *-updates repositories get replaced.