Processes os-prober output to stop false menu item
Prevents spurious items in the device selection menu by removing the leading
"/dev/" string (and EFI bootloader string) from the output of `os-prober` and
ensuring the device string gets correctly processed by the NVMe support
functions.

Also prints out the `os-prober` output to STDOUT (and thus written to log
file), which helps with any future debugging.

This regression occurred when NVMe support was integrated. Unit tests have not
been added in this commit, but will be done as part of #49.

Fixes #69