Retrieves memory bus width (32/64 bit) on launch
Retrieves whether Rescuezilla frontend is operating as a 32bit or 64bit
application on the initial launch of the application, rather than embedding
this information during Rescuezilla live environment ISO image build. This
decouples this value from the live environment build system, which is important
since Rescuezilla frontend may be packaged for other Linux distributions.

Also renames the variable memory_bus_width rather than register_width, in order
to be as clear and accurate as possible.

Also removes the ARCH variable (i386 or amd64), to simplify the information
presented to the user.