Disables Travis-CI shallow clone so build succeeds
Disables Travis-CI's git clone depth limit of 50 commits, as it seems to be
causing the build to fail during the deb package build, as the fact there was
50 commits since last git tag causes `git describe --tags` to return an error
(since the history is too shallow it can't calculate the number of commits
since last tag). The paths and filenames used by the construction of the deb
package contains that string, so the build fails.

Adds a TODO section to the Makefile as a reminder to leverage GNU Make 4.2's
"$(.SHELLSTATUS)" exit code feature when the build environment gets upgraded
and this version becomes available. When implemented, this will make similar
issues less cryptic.

Strangely, an earlier Pull Request successfully built and was merged, but a
subsequent build of the same commit failed. Examining the Travis-CI logs [1]
[2] suggested the difference in outcome may be attributable to the `docker
build` step, which downloads a different set of Ubuntu packages on the
Travis-CI builds that succeed versus the ones that don't (the step took 42
seconds on successful builds and 73 seconds on failed builds). However, the git
version appears the same between both versions, so it's still unclear why the
difference in outcome occurs. Also the reason why the `apt-get` command
downloads a slightly more complete package environment depending on some
Travis-CI invocations is also not clear.

That said, the change in git clone depth should fix the recent build
instability issue.

[1] https://travis-ci.org/github/rescuezilla/rescuezilla/builds/690094469

[2] https://travis-ci.org/github/rescuezilla/rescuezilla/builds/690064522