Removes restrictions for naming a backup
Removes the sed substitutions which prevent a user from entering characters
that are illegal for many filesystems.

Adds a simple sanity check to ensure the destination drive is writable with the
provided filename and cause a fatal error message otherwise. Clearly exiting
the application is not the ideal solution, but this usability problem will be
improved when the application is overhauled.

This assists with #15.