Checks if sfdisk, partclone.restore v1.0.4 exists
Adds an extra sanity check to generate a fatal dialog box if executables like
sfdisk and the v1.0.4 partclone.restore binary are missing. Both partclone and
sfdisk packages are dependencies of the Rescuezilla frontend so can be
reasonably assumed to be installed.

However, to ensure the ability to restore Redo Backup v1.0.4 backups, there are
two carefully named executables which provide the required older versions of
sfdisk and partclone.restore executables. Users who install Rescuezilla as a
deb package will not have these binaries, so it's very important to make this
situation clear to a user.

Removes a few localized print statements around where the modifications were
being done, as translated log files are not required.