Builds old sfdisk and partclone as git submodules
Builds a custom version of the sfdisk and partclone.restore executables on a
64-bit target platform to provide Redo Backup v1.0.4 backwards compatibility on
the AMD64 variant of Rescuezilla. The Rescuezilla AMD64 variant is based on
Ubuntu 20.04 Focal which does not have i386 support, so unlike Rescuezilla
i386, it cannot run the same binary from Redo Backup v1.0.4.

See #77 for a more detailed explanation of the issue.

Both the sfdisk and partclone projects needed changes to build on the Ubuntu
18.04 host environment. To manage these changes, each submodule points to a
fork hosted on Rescuezilla's GitHub organization.

The source code for sfdisk and partclone have been integrated into the
Rescuezilla repository as a git submodule. A static binary is built with the
Makefile also deploying both executables to the chroot directory. Notably, the
Makefile build disables the following filesystems: XFS, reiserfs (but not
reiser4) UFS, VMFS and JFS. See the comment in the Makefile for more detail.

Fixes #77