Detects and prints SMB stale file handle error
Prints an error message referring users to GitHub issue #84 when Rescuezilla is
unable to create a test file during an SMB connection. This informs users of a
bug when connecting to recent SMB servers (such as on Ubuntu 20.04).

I was unable to find a solution after a lot of debugging, so rather than
delaying Rescuezilla v1.0.6 for something that will only affect a handful of
advanced users, printing an error message linking to GitHub #84 is a reasonable
compromise.