Tweaks Rescuezilla wallpaper, icon, splash screen
Switches the Rescuezilla wallpaper and plymouth splash screen from being a
monochrome image of the title/mascot to being the actual 24-bit color version,
for improved aesthetics.  Resized the plymouth image to fit in 800x600 mode
without clipping.

Scales up the wallpaper to 1920x1080 (the pcmanfm file manager responsible for
the wallpaper display does an acceptable job of scaling the image down for
smaller display, and up for 4K).

Scaled the Rescuezilla mascot to be 2 pixels larger to completely fit the
rasterized icon. Also converted the forehead 'R' to be a path, so it correctly
displays on all machines, even without the Futura font installed.