Adds 'noprompt' to not prompt to eject media
Adds the 'noprompt' boot option to ensure that the "Please remove the
installation medium, then press ENTER" message is *not* displayed. Most
Rescuezilla users will boot using a key like F8/F12, rather changing the boot
options, so for most users the prompt is superfluous.

This improvement is of course not the root cause of #87, but it's a useful
improvement separate to fixing the root cause of that issue.