Limit systemd service timeout to 10s, not 90s
Limits the time systemd takes to start or stop a service to 10 seconds, as this
provides a better user experience when eg, a service is preventing shutdown.

This change is to improve the end-user experience as the Ubuntu 20.04 shutdown
is slowed down due to a bug in the PCManFM file manager [1]. Until this package
is fixed by Canonical, this workaround is an acceptable solution.

[1] https://bugs.launchpad.net/ubuntu/+source/pcmanfm/+bug/1878625

Fixes #87