Updates CHANGELOG, pkgs for Rescuezilla v1.0.6.1
Adds CHANGELOG entry for v1.0.6.1, and upgrades the debian package changelog
entries for the Rescuezilla frontend apps version 1.0.6.1-1, so that things
continue to build when this commit is tagged.

Note: internally, the build scripts that generate the deb package version using
`git describe --tags --abbrev=0`, which sets the version number. Because of
this, the build will fail if the deb package changelog does not reflect the
most recent git tag output from the `git describe` command above.

This means for every commit to successfully build without a single commit
failing, it is this very commit that must be tagged as 1.0.6.1.