Ensures partclone src dir not modified
Ensures two source files that are modified when building the old versions of
the partclone binary (that are used to provide Redo Backup v1.0.4 backwards
compatibility on 64-bit systems) maintain their original unmodified state.

This is an issue as the generated ISO images for Rescuezilla v1.0.6 from
TravisCI display with a "-dirty" version suffix indicating a working directory
which was not pristine. This source directory modification happens only during
a docker build on TravisCI, but strangely not in a local docker build.

For now, simply ensuring the two files are not modified when running the
subsequent build steps is an acceptable solution.

Fixes #102