Adds GVFS for PCManFM's smb://server/folder paths
Adds the backend packages for GNOME's userspace virtual filesystem. This allows
the PCManFM filemanager to resolve paths such as smb://fileserver/johnsmith
through a small window to enter the authentication details.

This provides an alternate graphical mechanism for users unfamiliar with the
Linux terminal to test whether their SMB/CIFS network share works, which is
useful for many users, eg #110.