Switches from GTK Bluebird theme to Breeze theme
Switches from the GTK2/3 theme called "Bluebird" to the GTK2/3 port of the
"Breeze" theme. The icon theme continues to use GNOME-Brave as it appears to be
more complete and provides better contrast to the desktop background color.

The lxappearance application was temporarily installed to graphically set the
widget and icon themes for both the 'ubuntu' and 'root' user accounts. Then the
~/.gtkrc-2.0 script and the ~/.config/gtk-3.0 directory were copied over and
committed.

Note: the original Breeze theme is the default theme of the KDE Plasma 5
desktop environment (which is based on the Qt library, not the GTK library).
Rescuezilla and the other desktop applications provided are GTK applications.

This change addresses aspects of task #52.