Renames Rescuezilla Perl script to rescuezilla.pl
Renames the /usr/sbin/rescuezilla script to /usr/sbin/rescuezilla.pl as a
temporary measure while Rescuezilla is ported from Python to Perl (task #111).

This allows the Perl script to be easily accessible while the Python port is
constructed. After the port is completed, the Perl version can be deleted.