Imports unmodified GParted v1.1.0 launcher script
Copies the "gparted.in" file from the official GParted git repository [1]
(specifically its v1.1.0 tag [2]) to be the basis of the Rescuezilla launch
script.

GParted installs this file into /usr/sbin/gparted path.

The script is released under GPLv2, which is fully compatible with GPLv3
("GParted is free software; you can redistribute it and/or modify it under the
terms of the GNU General Public License as published by the Free Software
Foundation; either version 2 of the License, or (at your option) any later
version.")

For maximum auditability, this commit does not make any changes any changes to
this file.

[1] https://gitlab.gnome.org/GNOME/gparted

[2] GPARTED_1_1_0