Replaces startmenu gear icon with lxpanel default
Removes the gear icon used originally by the adeskbar start menu on Redo Backup
and Recovery, that was kept for look-and-feel purposes when Rescuezilla
switched from adeskbar to lxpanel.

One user suggested they didn't realize Rescuezilla had a startmenu (and that
the gear icon was a settings launcher).

The default lxpanel monitor icon should be sufficient, and there's no longer
any need to imitate the look-and-feel of Redo Backup and Recovery.