Disables systemd-timesyncd to stop RTC changing
Masks out the systemd-timesyncd service, which provides Network Time Protocol
time synchronization because it modifies the hardware clock (also known as the
RTC) even with `hwclock --hctosys` being set at boot.

There's no simple way around modifying the hardware clock whenever the Linux
kernel synchronizes time: it happens within the kernel and the only way to
change this behavior is use a kernel compiled with a certain flag, which is not
feasible at this stage due to Secure Boot considerations.

[1] https://github.com/rescuezilla/rescuezilla/issues/107