Adds exploratory alpha "Image Explorer" mode
Adds a new mode to Rescuezilla to make it easier to explorer the contents of
backup images.

This feature uses the partclone's --restore_raw_file flag. This feature
unfortunately does *not* instantaneously mount a partclone backup image but
instead restores the backup image to a file. For larger images, this restore
operation takes a long time and uses a lot of space.

The Rescuezilla Image Explorer exploratory alpha sets the destination to a file
within the /tmp folder, so it's main memory (not disk space) that's the
limiting factor.

Given these limitations, Rescuezilla Image Explorer will *not* be broadly
useful until partclone has a sparse mount option. It still may be useful for
very small backup images. The feature will not be visible to users until these
limitations are resolved.

The magnifying glass image has been adapted from this public domain image
(Creative Commons Zero 1.0 License) [1]

https://github.com/rescuezilla/rescuezilla/issues/20

[1] https://openclipart.org/detail/304163/magnifying-glass