Integrate partclone-utils into build system
Integrates the build of partclone-utils into the build system for AMD64
targets.

The i386 build is not included because `partclone-util` depends on nbdkit,
which is not easily available on the Ubuntu 18.04 (Bionic) release that the
i386 build is based on.

Uses Debian's checkinstall to wrap the `make install` command into a deb
package for standardized ability to install/uninstall/upgrade the package using
dpkg and to a lesser extent, apt-get.