Stops considering failed geometry query as fatal
Stops considering the failure of sfdisk's querying of drive geometry as a fatal
error that should cause the backup operation to cease.

Users have reported issues around this on some hard drives [1].

Clonezilla does NOT check the exit code. The drive geometry file is for
informational purposes only and is not used, so there's little need to consider
this a fatal error.

[1] https://github.com/rescuezilla/rescuezilla/issues/122