Fixes restore of partial Redo backups
Fixes the ability to restore legacy Redo Backup and Recovery and Rescuezilla
v1.0.5 images when the backup image was not a whole disk but instead subset of
the partitions by generates warning when images are missing rather than
assuming their presence.

The legacy Redo Backup and Recovery format cannot distinguish between a failed
partclone backup and when a user chose not to backup a partition, a warning is
always printed. This also assists when Redo Backup and Recovery fails to backup
certain filesystems eg, minix filesystem.