Adds ability to cancel image folder scan
Adds the ability to cancel the image folder query by closing the popup. This
should begin to address user complaints that the image scanning takes too long.
If the user has 50+ images scanning might take a minutes, so the user can
simply cancel the image scan and use the Browse button to navigate directly to
their image. This means only 1 image needs to be scanned (which takes a few
seconds) rather than all of them.

Also updates the popup with information about the image that has just been
scanned, which will should also improve user experience.

[1] https://github.com/rescuezilla/rescuezilla/issues/126