Improves drive query please wait popup feedback
Provides more information to users watching the drive scan operation. This will
make it much more clear to a user what Rescuezilla is doing internally, which
should begin to address user complaints that the image scanning takes too long.
[1]

This commit does NOT add the ability to cancel the drive query operation.

[1] https://github.com/rescuezilla/rescuezilla/issues/126