Always restore all post-MBR gap, improve post-MBR
Always restores the entirety of the post-MBR gap (up to 1024 megabytes). In
Rescuezilla v2.0 a backup was made of the entire post-MBR gap (up-to 1024MB
just like Clonezilla) but during the restore only the first 256 kilobytes was
restored, due to the perceived complexity of calculating the offset to the
first partition on the destination disk based on just the sfdisk file. However
when the entire partition table is being overwritten it's not actually possible
for the post-MBR gap to accidentally overwrite a partition that the user wanted
to keep: the user can only choose from the partitions within the backup image.

Also detects when Clonezilla did not create a post-MBR gap (but instead created
hidden-data-after-mbr.notes.txt) so inform the user that Clonezilla is relying
on a GRUB re-install that as of writing is not yet supported by Rescuezilla.
[1]

Finally, improve how post-MBR gap files are globbed to better handle the
multi-disk case in the future.

These changes helps improve support and better match behavior with Clonezilla
[2]

[1] https://github.com/rescuezilla/rescuezilla/issues/131

[2] https://github.com/rescuezilla/rescuezilla/issues/146