Improves display of sfdisk errors
Improves the ability for the sfdisk partition table restore command to display
errors by combining stdout and stderr into a single stream, which makes it much
easier to see when sfdisk had an issue.

Also for legacy Redo Backup and Recovery images checks the exit code and
displays a non-fatal warning if it fails (for compatibility with louvetch's
unofficial Redo 2016 releases), rather than the earlier approach of ignoring
the exit code.

https://github.com/rescuezilla/rescuezilla/issues/147