Updates build instructions after targets renamed
Fixes the build instructions to make it easier for other developers to build
Rescuezilla.

The instructions were not kept up-to-date since renaming the Makefile targets
after the Ubuntu 20.10 (Groovy) build was added etc.

There's still a number of other changes that need to be implemented for
Rescuezilla to build on more recent builds of Ubuntu (see links in [1]).

[1] https://github.com/rescuezilla/rescuezilla/issues/161