Fixes Rescuezilla Python GUI Openbox automaximize
Ensures Openbox window manager is correctly configured to automaximize the
relatively new Rescuezilla Python GUI, which had updated the Openbox
configuration parameter "OB_APP_NAME" to "rescuezillapy" rather than
"Rescuezilla", as can be seen through the `obxprop` tool.

Also sets the title and type to ensures that dialog boxes do not get maximized.
For more information see the comments on the rc.xml file, and the output of
obxprop.

Issue #164