Add scripts to automate the creation of a virtualbox VM
Address #151 in part.

create_partitions_and_filesystems.sh must be called from the VM after
mounting the shared directory with this command:
    mkdir shared; mount -t vboxsf shared shared

squash! Add scripts to automate the creation of a virtualbox VM

Fix VM creation... The 2nd SATA disk was added while I was playing with the
VM and I didn't retry the script before committing.

squash! Add scripts to automate the creation of a virtualbox VM

Use sfdisk instead of hacking our way with fdisk.
Partition sdb with GPT. The 6th partition is of an unknown type.

https://github.com/rescuezilla/rescuezilla/issues/151