Upgrade partclone-nbd, partclone-utils to master
Switches the commit on the partclone-nbd submodule (used for Rescuezilla Image
Explorer) to point to the current latest (untagged) commit on the master branch
of the upstream repository [1], because the partclone-nbd author recently
resurfaced and merged my Pull Requests. No functional changes are made -- it's
effectively just a rebase.

Similarly switches the commit on the partclone-utils submodule (to be used by
Image Explorer in the future) to the current latest (untagged) commit on the
master branch of its upstream repository [1], because I have gained merge
privileges from the original author, and recently merged several Pull Requests.
This switch *does* include changes to functionality, but given its not used by
Image Explorer yet this is not a problem.

Note: now that both projects have maintainers, future development of
Rescuezilla intend to better track only tagged/versioned submodules rather than
my development branches. For now though, switching to the master branch of the
upstream projects should at least provide clearer tracking of changes.

[1] https://github.com/spaemann/partclone-nbd

[2] https://sourceforge.net/projects/partclone-utils/