Adds FOG/Redo Rescue/Foxclone/FSArchiver support
Adds the initial ability for Rescuezilla to restore and explore images created
by the other partclone-based applications: FOG Project, Redo Rescue and
Foxclone.

Also adds support for FSArchiver-based applications such as the (relatively
easy to use) command-line FSArchiver tool, and its qt-fsarchiver frontend.

The changes in this commit are NOT well tested as the code subject to a
refactor in a subsequent commit.

https://github.com/rescuezilla/rescuezilla/issues/194