Displays path *being* scanned, not already scanned
During the image folder query, switches to displaying the path of the image
that's currently scanning rather than the image that has just completed its
scan.

This makes it much more clear to the user which image is taking too long, and
therefore they can choose to move it to a subdirectory somewhere.

The 'Scanned' keyword doesn't have a translation string associated with it yet,
so no need to update this.