Improves drive query's handling of errors
Adds wrapper function to catch unexpected exceptions during the drive querying
thread, and a callback function to display the errors to the user (and return
to the Welcome screen).

This ensures the user is clearly informed of any exceptions. The same strategy
is used elsewhere in Rescuezilla but drive_query.py was never updated with this
approach.