Improves fast estimate of image size, filesystem
Improves the ability to estimate the size and filesystems of Clonezilla and
Redo images and does a minor refactor to make the logic more clear.

Specifically, for Clonezilla logical volume manage (LVM) logical volume (LVs)
and some md RAID disks did not have their size information available. And for
some versions of Redo and Clonezilla, the partition table is not available.

The best solution is to run partclone.info over all the partclone images, but
this is too slow to be used during the image query step where there could be
many tens of images being scanned. Also it wouldn't work for eg, dd images.  So
instead, when the estimate cannot be gleaned from the partition table, the
estimate can be generated by summing the file sizes of the split image files
(which may be compressed but this is ignored here). This provides some non-zero
number even if the image was created with partclone, partimage, ntfsclone, dd
etc. This number is evaluated relatively quickly, and gives the end-user a lot
of useful information, though confusing because it's approximately the used
space not the drive capacity.