Deletes obsolete `fsarchiver probe` parser
Removes the obsolete code and unit test used to parse the output of the
`fsarchiver probe` command, which lists all the drives and partitions in the
system. This command is why the original Redo Backup and Recovery tool used,
and therefore was used by Rescuezilla prior to v2.0.

`fsarchiver probe` doesn't have very good output (unable to produce exact byte
sizes etc), so for Rescuezilla v2.0 it made sense to running a bunch of
different tools and combining the results in exactly the way Clonezilla does
it.

The `fsarchiver probe` approach was ported over to Python as an exercise but
never used, so removing it can be safely removed.