Displays more info on PT refresh, LVM shutdown
Displays finer grain information about the commands used to refresh the
partition table, and that the LVM shutdown is occurring.

This should help user experience because the "Refreshing partition table"
message stays displayed on the screen for a relatively long time, and it's
possible for the LVM shutdown to take sometime.