Integrates xfce4-screenshooter screenshot app
Installs "xfce4-screenshooter" application and configures Openbox Window
Manager to bind the Print Screen key to launch the application.

The application allows the user to select the location in which to save the
image, and also makes it trivially easy to upload the image directly to Imgur,
which is a very nice feature when it comes to Rescuezilla user support.

The app is 900 kilobytes compressed, has few dependencies and is already
translated into a variety of languages.

Fixes #166