Adds ability to shutdown/reboot after completion
Adds the ability for the user to configure Rescuezilla to automatically
shutdown or reboot after a backup or restore operation has successfully
completed.

Integrates the Shutdown/Reboot string translations from the graphical-shutdown
script. Unfortunately GTK doesn't provide the ability to load multiple
translation domains, so duplicating the strings was the best solution.

Fixes #165