Resets net protocol combobox between operations
Ensures network protocol user-interface elements always remain consistent by
ensuring the Welcome page triggers the callbacks for the comboboxes, frames and
radio buttons.

This prevents the situation where after completing a backup operation using the
"Connected directly to my computer" radio button was selected but the "Shared
over the network" frame was displayed.

This commit resets the combobox to the initial state, it does NOT achieve the
ideal state where changing the protocol on the Backup operation would also set
the protocol on the Restore or Image Explorer modes.

https://github.com/rescuezilla/rescuezilla/issues/81

https://github.com/rescuezilla/rescuezilla/issues/118