Grows filesystem to fit partition size
Always try to grow the filesystem to fit the partition size, which is the
behavior of Clonezilla's "-r" option used by default in its restoredisk and
restoreparts function. Installs missing fatresize package as the other packages
listed in [1] are already installed.

Some filesystems (XFS, JFS, BTRFS and NILFS) requires mounting before resizing,
and some (but not all) do a filesystem check step first. This commit ports
Clonezilla's ocs-resize-part script to Python, so the exact approach used in
Clonezilla is replicated here.

[1] https://github.com/rescuezilla/rescuezilla/issues/18#issuecomment-823057082