Updates translation skeleton inc Image Explorer
Updates the Python source code, and existing skeleton files for all translated
files to ensure the translation files reflect the current state of Rescuezilla.

This includes new features such as Clonezilla's ability to reinstall GRUB
bootloader and adding a translation skeleton for the parts of Image Explorer
(beta) feature that are unlikely to change.