Displays data being saved in central section
Displays the Clonezilla metadata that's being saved in the central
informational section of the backup progress page, in addition to the status
bar that it has always been display in.

This should improve the user experience at the start of the backup when it's
not clear that Rescuezilla is doing anything.

Also does a gettext lookup to the line that's being displayed, but does not
update the translation templates in this commit.