Sets 'Description' field to 100 px default width
Sets the image 'Description' column (which contains the user-provided
description text) to a default width of 100 pixels (with the column still
resizable) because the content of the Description field for some image types
(like QtFSArchiver) often has very wide lines.

The fixed size ensures the much more important 'Partition' column is always
visible.