Sets C locale before parsing timestamp strings
Sets the C (POSIX) locale temporary when parsing the timestamp strings which
are part of the Foxclone, FSArchiver, Redo Rescue image formats.

This is required because strptime command interprets fields (eg, day of the
week, "Mon") using the relevant localized version, which causes an exception
for non-English locales.  Since those image formats produce timestamp strings
always in English, the simplest solution is to use the C locale.