Displays error when restoring to smaller disk
Displays error message when restoring an image where the final partition is
beyond the end of the destination disk capacity, based on analyzing the image's
sfdisk partition table and comparing it to the destination disk.

This comparison is likely off by a few bytes for GPT disks that has a secondary
GPT backup at the end of the disk, but it should be very useful for a lot of
users.

[1] https://github.com/rescuezilla/rescuezilla/wiki/HOWTO:-Restoring-to-a-smaller-disk.-Eg,-1000GB-HDD-to-500GB-SSD