Polls nbd-client association until ready
Retries Image Explorer's nbd-client associations for the join and decompress
steps, which sometimes fails with "Error: Socket failed: Connection refused".
By retrying every 1 second (with a 10 second timeout), the incidence of this
error appears to have been removed.

The retry function was being used in the qemu-nbd case, and has been moved to
its own function to be more broadly useful.