Adds 'when ready click next' to partition config
Adds a "When you are ready to continue, click Next" to the pages that ask users
to enable/disable the partitions to backup or map partitions to restore.

This should reduce the confusion that some users have on these pages.

Also slightly more descriptive Glade GTKBuilder id values for such messages.