Improves double-clicking partition mapping table
Toggles the partition when double-clicking on a row on the backup mode's
partition to backup screen, rather than going to the next screen.

Also adds similar toggle behavior to the restore mode's partition mapping
screen (though that didn't have the behavior of going to the next tab when
double-clicking).

This should improve usability on two screens that are already Rescuezilla's
most complex.