Adds "Cloning" mode, ability to restore VM images
Adds a new "Cloning" mode for device-to-device operation, and adds the ability
for the existing "Restore" mode to work with QemuImage VM images (rather than
only being able to process them using Image Explorer)

The QemuImage class has been separated into a MetadataOnlyImage superclass to
assist with this.

The Cloning operation is effectively a backup followed by a restore, so this
change touches a lot of the code base, so was most effectively developed as a
very monolithic commit.

The drive icon used by the backup, restore and now clone logo images has been
adapted from the oxygen-icon-theme's drive-harddisk.png file [1] [2]

[1] https://packages.debian.org/buster/oxygen-icon-theme

[2] /usr/share/icons/oxygen/base/256x256/devices/drive-harddisk.png

https://github.com/rescuezilla/rescuezilla/issues/47

Fixes #47