Associates VM images as RW, NOT as read-only
Associates VM images with a device node using qemu-nbd not as read-only, so
that during a backup operation NTFS partitions can be modified using 'ntfsfix'
to remove the very common issue where Windows has been hibernated, causing the
partclone operation to fail.

It's not ideal for the source image to be exposed as writeable, ideally they'll
be a NBD copy-on-write layer, but qemu-nbd doesn't seem to support this.