Runs partclone.info when no PT in Clonezilla image
Scans the Clonezilla image's sole (likely partclone) image when the Clonezilla
image doesn't have an sfdisk partition table, such as when Clonezilla's
saveparts function has been used on a disk without a partition table (like many
RAID disks).

It's not guaranteed that the filesystem image was made by partclone (rather
than being eg, an dd image), but using partclone.info allows an more accurate
filesystem size estimate to be determined, which is very useful to display to the
end user.

https://github.com/rescuezilla/rescuezilla/issues/43