Prevents 'sfdisk missing' when no MBR
Prevents displaying sfdisk's "extended partitions are missing" message when the
image reports it does not have a partition table (almost always the existence
of an MBR file).

This is the case for Clonezilla which can run saveparts on disks without
partition tables (filesystem directly on disk) and thus correctly have no MBR
(or sfdisk backup). Some Clonezilla images have an MBR file and thus a
partition table but have an empty sfdisk file, so the message should be
displayed when the image has a partition table.

https://github.com/rescuezilla/rescuezilla/issues/43