Populates test sfdict, parted dicts correctly
Populates the unit test dictionaries for sfdict and parted correctly, as they
were being clobbered each time.

The "unit test" is not asserted anything as the CombinedDriveState approach
that Clonezilla uses (and Rescuezilla ports) may change, but the "unit  test"
is still very useful as a quick sandbox testing environment.