Prevents backup/clone of RAID *member* drives
Improves the user-interface to stop a user trying to backup or clone RAID
member drives, because the actual Linux RAID device such as /dev/md0 or
/dev/md127 works.

Exposes a boolean in the CombinedDriveState structure for a given drive so it's
easy to check if there is a partition that has a RAID Linux member filesystem.

Also hides RAID member drives behind the "Show hidden devices (for advance
users)" checkbox, and additionally exposes the standard Linux RAID devices such
as /dev/md0 rather than the previous approach of hiding them.

https://github.com/rescuezilla/rescuezilla/issues/43

https://github.com/rescuezilla/rescuezilla/issues/174